Add files via upload
Here is a little script to transcribe and generate a train.txt file as the toirtoise model trainer expect it to be. I have done it because, I download a supper good quality voice dataset wich is structured with speaker directory and short audio sample of high quality.

A csv list all of the speaker, number of file, duration in mn , f0_mean. I use this csv to generate the iteration as the audio_file have all the same format "speaker_num_audio_file.wav". One thing I wonder is that the process speed isn't so great. As i have only one gpu I didn't optimize by multi-threading the generation model of whisperx